requirements.txt
absl-py==2.1.0
ale-py==0.11.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
bleach==6.3.0
blinker==1.9.0
box2d-py==2.3.5
bs4==0.0.2
cachetools==5.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
chex==0.1.89
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Cython==0.29.37
decorator==5.2.1
distlib==0.3.9
Django==5.1.3
docutils==0.21.2
etils==1.12.2
exceptiongroup==1.2.2
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.16.1
Flask==3.1.1
flatbuffers==24.3.25
flax==0.10.6
fonttools==4.54.1
fsspec==2025.3.2
gast==0.6.0
glfw==2.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
gymnasium==1.1.1
h11==0.16.0
h5py==3.12.1
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface_hub==1.1.2
humanize==4.12.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
joblib==1.4.2
kaggle==1.7.4.5
kagglehub==0.3.4
keras==3.7.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matlib==0.1.2
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.1
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
mujoco==3.3.1
mujoco-py==2.1.2.14
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.6
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optax==0.2.4
optree==0.13.1
orbax-checkpoint==0.11.12
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.6.0
proglog==0.1.11
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.18.0
PyMySQL==1.1.1
PyOpenGL==3.1.9
pyparsing==3.1.4
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9.1
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
selenium==4.26.1
shellingham==1.5.4
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
sqlparse==0.5.1
swig==4.3.1
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
tensorstore==0.1.74
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tomli==2.2.1
toolz==1.0.0
torch==2.7.0
tqdm==4.67.0
treescope==0.1.9
trio==0.27.0
trio-websocket==0.11.1
typer-slim==0.20.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.2.3
virtualenv==20.27.1
virtualenvwrapper-win==1.2.7
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.0
wsproto==1.2.0
zipp==3.21.0
